separate components and use error handling from auth0
